Fix build

+++ Web Sites/TVM/adm_media_files_ppvmodules.aspx.cs
@@ -317,7 +317,7 @@ public partial class adm_media_files_ppvmodules : System.Web.UI.Page
                 // check if changes in the start date require future index update call, incase updatedStartDate is in more than 2 years we don't update the index (per Ira's request)
                 if (RabbitHelper.IsFutureIndexUpdate(prevStartDate, updatedStartDate))
                 {
-                    if (!RabbitHelper.InsertFreeItemsIndexUpdate(nLogedInGroupID, ApiObjects.eObjectType.Media, new List<int>() { mediaID }, updatedStartDate.Value))
+                    if (!RabbitHelper.InsertFreeItemsIndexUpdate(nLogedInGroupID, ApiObjects.eObjectType.Media, new List<long>() { mediaID }, updatedStartDate.Value))
                     {
                         log.Error(string.Format("Failed inserting free items index update for startDate: {0}, mediaID: {1}, groupID: {2}", updatedStartDate.Value, mediaID, nLogedInGroupID));
                     }
@@ -326,7 +326,7 @@ public partial class adm_media_files_ppvmodules : System.Web.UI.Page
                 // check if changes in the end date require future index update call, incase updatedEndDate is in more than 2 years we don't update the index (per Ira's request)
                 if (RabbitHelper.IsFutureIndexUpdate(prevEndDate, updatedEndDate))
                 {
-                    if (!RabbitHelper.InsertFreeItemsIndexUpdate(nLogedInGroupID, ApiObjects.eObjectType.Media, new List<int>() { mediaID }, updatedEndDate.Value))
+                    if (!RabbitHelper.InsertFreeItemsIndexUpdate(nLogedInGroupID, ApiObjects.eObjectType.Media, new List<long>() { mediaID }, updatedEndDate.Value))
                     {
                         log.Error(string.Format("Failed inserting free items index update for endDate: {0}, mediaID: {1}, groupID: {2}", updatedEndDate.Value, mediaID, nLogedInGroupID));
                     }
